Created a BasicMovement script
Branch for the Basic Movement in Frogger. This script handles the basic movement allowing the player to hop in all 4 directions using keyboard input.

This is my first commit, I just redid everything to link my Github and Unity project so I am hoping this doesnt mess anything up for the group. The script for basic movement is called BasicMovement.cs.